Move local models to new folder
