+++ timeseries/src/autogluon/timeseries/models/local/__init__.py
@@ -1,3 +1,8 @@
 from .naive import NaiveModel, SeasonalNaiveModel
-from .statsforecast import AutoARIMAModel, AutoETSModel, DynamicOptimizedThetaModel
-from .statsmodels import ARIMAModel, ETSModel, ThetaModel
+from .statsmodels import ETSStatsmodelsModel, ARIMAStatsmodelsModel, ThetaStatsmodelsModel
+from .statsforecast import AutoARIMAModel, AutoETSModel, DynamicOptimizedThetaModel, ThetaModel, ETSModel
+
+
+import joblib.externals.loky
+
+joblib.externals.loky.process_executor._MAX_MEMORY_LEAK_SIZE = int(3e10)
