+++ timeseries/src/autogluon/timeseries/models/local/abstract_local_model.py
@@ -1,38 +1,42 @@
-import hashlib
 import logging
-from multiprocessing import cpu_count
+import time
+from multiprocessing import cpu_count, TimeoutError
 from typing import Any, Dict, List, Optional, Union
+from autogluon.timeseries.dataset import TimeSeriesDataFrame
 
+import numpy as np
 import pandas as pd
 from joblib import Parallel, delayed
+from scipy.stats import norm
 
-from autogluon.timeseries.dataset.ts_dataframe import ITEMID, TIMESTAMP, TimeSeriesDataFrame
+from autogluon.core.utils.exceptions import TimeLimitExceeded
+from autogluon.timeseries.dataset.ts_dataframe import ITEMID, TimeSeriesDataFrame
 from autogluon.timeseries.models.abstract import AbstractTimeSeriesModel
+from autogluon.timeseries.utils.forecast import get_forecast_horizon_index_ts_dataframe
 from autogluon.timeseries.utils.seasonality import get_seasonality
-from autogluon.timeseries.utils.warning_filters import statsmodels_joblib_warning_filter
+from autogluon.timeseries.utils.warning_filters import statsmodels_joblib_warning_filter, statsmodels_warning_filter
 
 logger = logging.getLogger(__name__)
 
 
-def hash_ts_dataframe_items(ts_dataframe: TimeSeriesDataFrame) -> pd.Series:
-    """Hash each time series in the dataset to a 32-character hex string.
-
-    Hash is computed based on the timestamps and values of the time series (item_id is ignored).
-
-    This means that any model that doesn't use static features will make identical predictions for two time series
-    with the same hash value (assuming no collisions).
-    """
-    df_with_timestamp = ts_dataframe.reset_index(level=TIMESTAMP)
-    hash_per_timestep = pd.util.hash_pandas_object(df_with_timestamp, index=False)
-    # groupby preserves the order of the timesteps
-    return hash_per_timestep.groupby(level=ITEMID, sort=False).apply(lambda x: hashlib.md5(x.values).hexdigest())
-
-
 class AbstractLocalModel(AbstractTimeSeriesModel):
+    """Abstract class for local forecasting models that are trained separately for each time series.
+
+    Prediction is parallelized across CPU cores using joblib.Parallel.
+
+    Attributes
+    ----------
+    allowed_local_model_args : List[str]
+        Argument that can be passed to the underlying local model.
+    max_ts_length : int, optional
+        If not None, only the last ``max_ts_length`` time steps of each time series will be used to train the model.
+    default_n_jobs : Union[int, float]
+        Default number of CPU cores used to train models. If float, this fraction of CPU cores will be used.
+    """
+
     allowed_local_model_args: List[str] = []
-    # Use 50% of the cores since some models rely on parallel ops and are actually slower if n_jobs=-1
-    DEFAULT_N_JOBS: Union[float, int] = 0.5
-    MAX_TS_LENGTH: Optional[int] = None
+    max_ts_length: Optional[int] = None
+    default_n_jobs: Union[int, float] = 0.5
 
     def __init__(
         self,
@@ -56,7 +60,7 @@ class AbstractLocalModel(AbstractTimeSeriesModel):
         if hyperparameters is None:
             hyperparameters = {}
         # TODO: Replace with 'num_cpus' argument passed to fit (after predictor API is changed)
-        n_jobs = hyperparameters.get("n_jobs", self.DEFAULT_N_JOBS)
+        n_jobs = hyperparameters.get("n_jobs", self.default_n_jobs)
         if isinstance(n_jobs, float) and 0 < n_jobs <= 1:
             self.n_jobs = max(int(cpu_count() * n_jobs), 1)
         elif isinstance(n_jobs, int):
@@ -64,7 +68,8 @@ class AbstractLocalModel(AbstractTimeSeriesModel):
         else:
             raise ValueError(f"n_jobs must be a float between 0 and 1 or an integer (received n_jobs = {n_jobs})")
         self._local_model_args: Dict[str, Any] = None
-        self._cached_predictions: Dict[str, pd.DataFrame] = {}
+        self._seasonal_period: Optional[int] = None
+        self.time_limit: Optional[float] = None
 
     def _fit(self, train_data: TimeSeriesDataFrame, time_limit: int = None, **kwargs):
         self._check_fit_params()
@@ -88,69 +93,99 @@ class AbstractLocalModel(AbstractTimeSeriesModel):
 
         if "seasonal_period" not in local_model_args or local_model_args["seasonal_period"] is None:
             local_model_args["seasonal_period"] = get_seasonality(train_data.freq)
-        self.freq = train_data.freq
+        self._seasonal_period = local_model_args["seasonal_period"]
 
-        self._local_model_args = self._update_local_model_args(local_model_args=local_model_args, data=train_data)
+        self._local_model_args = self._update_local_model_args(local_model_args=local_model_args)
+        self.time_limit = time_limit
 
         logger.debug(f"{self.name} is a local model, so the model will be fit at prediction time.")
         return self
 
-    def _update_local_model_args(
-        self, local_model_args: Dict[str, Any], data: TimeSeriesDataFrame, **kwargs
-    ) -> Dict[str, Any]:
+    def _update_local_model_args(self, local_model_args: Dict[str, Any]) -> Dict[str, Any]:
         return local_model_args
 
-    def predict(self, data: TimeSeriesDataFrame, quantile_levels: List[float] = None, **kwargs) -> TimeSeriesDataFrame:
-        if quantile_levels is None:
-            quantile_levels = self.quantile_levels
+    def predict(self, data: TimeSeriesDataFrame, **kwargs) -> TimeSeriesDataFrame:
+        if self.max_ts_length is not None:
+            logger.debug(f"Shortening all time series to at most {self.max_ts_length}")
+            data = data.groupby(level=ITEMID, sort=False).tail(self.max_ts_length)
 
-        if self.freq != data.freq:
-            raise RuntimeError(
-                f"{self.name} has frequency '{self.freq}', which doesn't match the frequency "
-                f"of the dataset '{data.freq}'."
-            )
-        if self.MAX_TS_LENGTH is not None:
-            logger.debug(f"Shortening all time series to at most {self.MAX_TS_LENGTH}")
-            data = data.groupby(level=ITEMID, sort=False).tail(self.MAX_TS_LENGTH)
+        df = pd.DataFrame(data).reset_index(level=ITEMID)
+        all_series = (ts for _, ts in df.groupby(by=ITEMID, as_index=False, sort=False)[self.target])
 
-        self._fit_and_cache_predictions(data, quantile_levels=quantile_levels)
-        item_id_to_prediction = {}
-        for item_id, ts_hash in hash_ts_dataframe_items(data).items():
-            item_id_to_prediction[item_id] = self._cached_predictions[ts_hash]
-        predictions_df = pd.concat(item_id_to_prediction)
-        predictions_df.index.rename([ITEMID, TIMESTAMP], inplace=True)
+        # timeout ensures that no individual job takes longer than time_limit
+        # TODO: a job started late may still exceed time_limit - how to prevent that?
+        timeout = None if self.n_jobs == 1 else self.time_limit
+        # end_time ensures that no new jobs are started after time_limit is exceeded
+        end_time = None if self.time_limit is None else time.time() + self.time_limit
+        executor = Parallel(self.n_jobs, timeout=timeout)
+
+        try:
+            with statsmodels_joblib_warning_filter(), statsmodels_warning_filter():
+                predictions = executor(delayed(self._predict_wrapper)(ts, end_time=end_time) for ts in all_series)
+        except TimeoutError:
+            raise TimeLimitExceeded
+
+        predictions_df = pd.concat(predictions)
+        predictions_df.index = get_forecast_horizon_index_ts_dataframe(data, self.prediction_length)
         return TimeSeriesDataFrame(predictions_df)
 
-    def _fit_and_cache_predictions(self, data: TimeSeriesDataFrame, quantile_levels: List[float]):
-        data_hash = hash_ts_dataframe_items(data)
-        items_to_fit = [item_id for item_id, ts_hash in data_hash.items() if ts_hash not in self._cached_predictions]
-        if len(items_to_fit) > 0:
-            logger.debug(f"{self.name} received {len(items_to_fit)} new items to predict, generating predictions")
-            target_series = data[self.target]
-            time_series_to_fit = (target_series.loc[item_id] for item_id in items_to_fit)
-            with statsmodels_joblib_warning_filter():
-                predictions = Parallel(n_jobs=self.n_jobs)(
-                    delayed(self._predict_with_local_model)(
-                        time_series=ts,
-                        prediction_length=self.prediction_length,
-                        freq=self.freq,
-                        quantile_levels=quantile_levels,
-                        local_model_args=self._local_model_args.copy(),
-                    )
-                    for ts in time_series_to_fit
-                )
-            for item_id, preds in zip(items_to_fit, predictions):
-                self._cached_predictions[data_hash.loc[item_id]] = preds
-            # Make sure cached predictions can be reused by other models
-            self.save()
+    def score_and_cache_oof(
+        self, val_data: TimeSeriesDataFrame, store_val_score: bool = False, store_predict_time: bool = False
+    ) -> None:
+        super().score_and_cache_oof(val_data, store_val_score, store_predict_time)
+        # Remove time_limit for future predictions
+        self.time_limit = None
 
-    @staticmethod
+    def _predict_wrapper(self, time_series: pd.Series, end_time: Optional[float] = None) -> pd.DataFrame:
+        if end_time is not None and time.time() >= end_time:
+            raise TimeLimitExceeded
+        try:
+            result = self._predict_with_local_model(
+                time_series=time_series,
+                freq=self.freq,
+                prediction_length=self.prediction_length,
+                quantile_levels=self.quantile_levels,
+                local_model_args=self._local_model_args.copy(),
+            )
+        except Exception as e:
+            result = seasonal_naive_forecast(
+                target=time_series.values.ravel(),
+                prediction_length=self.prediction_length,
+                quantile_levels=self.quantile_levels,
+                seasonal_period=self._seasonal_period,
+            )
+        return result
+
     def _predict_with_local_model(
+        self,
         time_series: pd.Series,
-        freq: str,
-        prediction_length: int,
-        quantile_levels: List[float],
         local_model_args: dict,
-        **kwargs,
     ) -> pd.DataFrame:
         raise NotImplementedError
+
+
+def seasonal_naive_forecast(
+    target: np.ndarray, prediction_length: int, quantile_levels: List[float], seasonal_period: int
+) -> pd.DataFrame:
+    """Generate seasonal naive forecast, predicting the last observed value from the same period."""
+    forecast = {}
+    if len(target) > seasonal_period and seasonal_period > 1:
+        indices = [len(target) - seasonal_period + k % seasonal_period for k in range(prediction_length)]
+        forecast["mean"] = target[indices]
+        residuals = target[seasonal_period:] - target[:-seasonal_period]
+
+        sigma = np.sqrt(np.mean(np.square(residuals)))
+        num_full_seasons = np.arange(1, prediction_length + 1) // seasonal_period
+        sigma_per_timestep = sigma * np.sqrt(num_full_seasons + 1)
+    else:
+        # Fall back to naive forecast
+        forecast["mean"] = np.full(shape=[prediction_length], fill_value=target[-1])
+        residuals = target[1:] - target[:-1]
+
+        sigma = np.sqrt(np.mean(np.square(residuals)))
+        sigma_per_timestep = sigma * np.sqrt(np.arange(1, prediction_length + 1))
+
+    for q in quantile_levels:
+        forecast[str(q)] = forecast["mean"] + norm.ppf(q) * sigma_per_timestep
+
+    return pd.DataFrame(forecast)
